refactor: pass variable limit to fetch document methods
