+++ packages/rs-drive/src/drive/identity/withdrawals/document/fetch_oldest_withdrawal_documents_by_status/v0/mod.rs
@@ -1,6 +1,6 @@
 use crate::drive::document::query::QueryDocumentsOutcomeV0Methods;
-use crate::drive::identity::withdrawals::WithdrawalTransactionIndex;
 use crate::drive::Drive;
+use crate::error::drive::DriveError;
 use crate::error::Error;
 use crate::query::{DriveQuery, InternalClauses, OrderClause, WhereClause};
 use dpp::data_contract::accessors::v0::DataContractV0Getters;
@@ -14,60 +14,58 @@ use platform_version::version::PlatformVersion;
 use std::collections::BTreeMap;
 
 impl Drive {
-    // TODO(withdrawals): Currently it queries only up to 100 documents.
-    //  It works while we don't have pooling
-
-    pub(super) fn find_up_to_100_withdrawal_documents_by_status_and_transaction_indices_v0(
+    pub(super) fn fetch_oldest_withdrawal_documents_by_status_v0(
         &self,
-        status: withdrawals_contract::WithdrawalStatus,
-        transaction_indices: &[WithdrawalTransactionIndex],
+        status: u8,
+        limit: u16,
         transaction: TransactionArg,
         platform_version: &PlatformVersion,
     ) -> Result<Vec<Document>, Error> {
+        let data_contract_id = withdrawals_contract::ID;
+
+        let contract_fetch_info = self
+            .get_contract_with_fetch_info_and_fee(
+                data_contract_id.to_buffer(),
+                None,
+                true,
+                transaction,
+                platform_version,
+            )?
+            .1
+            .ok_or_else(|| {
+                Error::Drive(DriveError::CorruptedCodeExecution(
+                    "Can't fetch data contract",
+                ))
+            })?;
+
+        let document_type = contract_fetch_info
+            .contract
+            .document_type_for_name(withdrawal::NAME)?;
+
         let mut where_clauses = BTreeMap::new();
 
+        //todo: make this lazy loaded or const
         where_clauses.insert(
             withdrawal::properties::STATUS.to_string(),
             WhereClause {
                 field: withdrawal::properties::STATUS.to_string(),
                 operator: crate::query::WhereOperator::Equal,
-                value: Value::U8(status as u8),
-            },
-        );
-
-        where_clauses.insert(
-            withdrawal::properties::TRANSACTION_INDEX.to_string(),
-            WhereClause {
-                field: withdrawal::properties::TRANSACTION_INDEX.to_string(),
-                operator: crate::query::WhereOperator::In,
-                value: Value::Array(
-                    transaction_indices
-                        .iter()
-                        .map(|index| Value::U64(*index))
-                        .collect::<Vec<_>>(),
-                ),
+                value: Value::U8(status),
             },
         );
 
         let mut order_by = IndexMap::new();
 
         order_by.insert(
-            withdrawal::properties::TRANSACTION_INDEX.to_string(),
+            withdrawal::properties::UPDATED_AT.to_string(),
             OrderClause {
-                field: withdrawal::properties::TRANSACTION_INDEX.to_string(),
+                field: withdrawal::properties::UPDATED_AT.to_string(),
                 ascending: true,
             },
         );
 
-        let cache = self.cache.read().unwrap();
-
-        let document_type = cache
-            .system_data_contracts
-            .withdrawals
-            .document_type_for_name(withdrawal::NAME)?;
-
         let drive_query = DriveQuery {
-            contract: &cache.system_data_contracts.withdrawals,
+            contract: &contract_fetch_info.contract,
             document_type,
             internal_clauses: InternalClauses {
                 primary_key_in_clause: None,
@@ -77,7 +75,7 @@ impl Drive {
                 equal_clauses: where_clauses,
             },
             offset: None,
-            limit: None,
+            limit: Some(limit),
             order_by,
             start_at: None,
             start_at_included: false,
@@ -98,7 +96,8 @@ impl Drive {
 
 #[cfg(test)]
 mod tests {
-    use crate::drive::identity::withdrawals::WithdrawalTransactionIndex;
+    use super::*;
+    use crate::drive::config::DEFAULT_QUERY_LIMIT;
     use crate::tests::helpers::setup::{
         setup_document, setup_drive_with_initial_state_structure, setup_system_data_contract,
     };
@@ -112,10 +111,8 @@ mod tests {
     use dpp::version::PlatformVersion;
     use dpp::withdrawal::Pooling;
 
-    use super::*;
-
     #[test]
-    fn test_find_pooled_withdrawal_documents_by_transaction_index() {
+    fn test_return_list_of_documents() {
         let drive = setup_drive_with_initial_state_structure();
 
         let transaction = drive.grove.start_transaction();
@@ -128,9 +125,18 @@ mod tests {
 
         setup_system_data_contract(&drive, &data_contract, Some(&transaction));
 
-        let owner_id = Identifier::new([1u8; 32]);
+        let documents = drive
+            .fetch_oldest_withdrawal_documents_by_status(
+                withdrawals_contract::WithdrawalStatus::QUEUED.into(),
+                DEFAULT_QUERY_LIMIT,
+                Some(&transaction),
+                platform_version,
+            )
+            .expect("to fetch documents by status");
 
-        let transaction_index: WithdrawalTransactionIndex = 1;
+        assert_eq!(documents.len(), 0);
+
+        let owner_id = Identifier::new([1u8; 32]);
 
         let document = get_withdrawal_document_fixture(
             &data_contract,
@@ -140,13 +146,13 @@ mod tests {
                 "coreFeePerByte": 1u32,
                 "pooling": Pooling::Never as u8,
                 "outputScript": CoreScript::from_bytes((0..23).collect::<Vec<u8>>()),
-                "status": withdrawals_contract::WithdrawalStatus::POOLED as u8,
-                "transactionIndex": transaction_index,
+                "status": withdrawals_contract::WithdrawalStatus::QUEUED as u8,
+                "transactionIndex": 1u64,
             }),
             None,
             platform_version.protocol_version,
         )
-        .expect("expected to get withdrawal document");
+        .expect("expected withdrawal document");
 
         let document_type = data_contract
             .document_type_for_name(withdrawal::NAME)
@@ -160,15 +166,50 @@ mod tests {
             Some(&transaction),
         );
 
-        let found_document = drive
-            .find_up_to_100_withdrawal_documents_by_status_and_transaction_indices(
-                withdrawals_contract::WithdrawalStatus::POOLED,
-                &[transaction_index],
+        let document = get_withdrawal_document_fixture(
+            &data_contract,
+            owner_id,
+            platform_value!({
+                "amount": 1000u64,
+                "coreFeePerByte": 1u32,
+                "pooling": Pooling::Never as u8,
+                "outputScript": CoreScript::from_bytes((0..23).collect::<Vec<u8>>()),
+                "status": withdrawals_contract::WithdrawalStatus::POOLED,
+                "transactionIndex": 2u64,
+            }),
+            None,
+            platform_version.protocol_version,
+        )
+        .expect("expected withdrawal document");
+
+        setup_document(
+            &drive,
+            &document,
+            &data_contract,
+            document_type,
+            Some(&transaction),
+        );
+
+        let documents = drive
+            .fetch_oldest_withdrawal_documents_by_status(
+                withdrawals_contract::WithdrawalStatus::QUEUED.into(),
+                DEFAULT_QUERY_LIMIT,
                 Some(&transaction),
                 platform_version,
             )
-            .expect("to find document by it's transaction id");
+            .expect("to fetch documents by status");
 
-        assert_eq!(found_document.len(), 1);
+        assert_eq!(documents.len(), 1);
+
+        let documents = drive
+            .fetch_oldest_withdrawal_documents_by_status(
+                withdrawals_contract::WithdrawalStatus::POOLED.into(),
+                DEFAULT_QUERY_LIMIT,
+                Some(&transaction),
+                platform_version,
+            )
+            .expect("to fetch documents by status");
+
+        assert_eq!(documents.len(), 1);
     }
 }
