+++ packages/rs-drive/src/drive/identity/withdrawals/document/find_withdrawal_documents_by_status_and_transaction_indices/v0/mod.rs
@@ -1,6 +1,6 @@
 use crate::drive::document::query::QueryDocumentsOutcomeV0Methods;
+use crate::drive::identity::withdrawals::WithdrawalTransactionIndex;
 use crate::drive::Drive;
-use crate::error::drive::DriveError;
 use crate::error::Error;
 use crate::query::{DriveQuery, InternalClauses, OrderClause, WhereClause};
 use dpp::data_contract::accessors::v0::DataContractV0Getters;
@@ -14,57 +14,61 @@ use platform_version::version::PlatformVersion;
 use std::collections::BTreeMap;
 
 impl Drive {
-    pub(super) fn fetch_up_to_100_oldest_withdrawal_documents_by_status_v0(
+    // TODO(withdrawals): Currently it queries only up to 100 documents.
+    //  It works while we don't have pooling
+
+    pub(super) fn find_withdrawal_documents_by_status_and_transaction_indices_v0(
         &self,
-        status: u8,
+        status: withdrawals_contract::WithdrawalStatus,
+        transaction_indices: &[WithdrawalTransactionIndex],
+        limit: u16,
         transaction: TransactionArg,
         platform_version: &PlatformVersion,
     ) -> Result<Vec<Document>, Error> {
-        let data_contract_id = withdrawals_contract::ID;
-
-        let contract_fetch_info = self
-            .get_contract_with_fetch_info_and_fee(
-                data_contract_id.to_buffer(),
-                None,
-                true,
-                transaction,
-                platform_version,
-            )?
-            .1
-            .ok_or_else(|| {
-                Error::Drive(DriveError::CorruptedCodeExecution(
-                    "Can't fetch data contract",
-                ))
-            })?;
-
-        let document_type = contract_fetch_info
-            .contract
-            .document_type_for_name(withdrawal::NAME)?;
-
         let mut where_clauses = BTreeMap::new();
 
-        //todo: make this lazy loaded or const
         where_clauses.insert(
             withdrawal::properties::STATUS.to_string(),
             WhereClause {
                 field: withdrawal::properties::STATUS.to_string(),
                 operator: crate::query::WhereOperator::Equal,
-                value: Value::U8(status),
+                value: Value::U8(status as u8),
+            },
+        );
+
+        where_clauses.insert(
+            withdrawal::properties::TRANSACTION_INDEX.to_string(),
+            WhereClause {
+                field: withdrawal::properties::TRANSACTION_INDEX.to_string(),
+                operator: crate::query::WhereOperator::In,
+                value: Value::Array(
+                    transaction_indices
+                        .iter()
+                        .map(|index| Value::U64(*index))
+                        .collect::<Vec<_>>(),
+                ),
             },
         );
 
         let mut order_by = IndexMap::new();
 
         order_by.insert(
-            withdrawal::properties::UPDATED_AT.to_string(),
+            withdrawal::properties::TRANSACTION_INDEX.to_string(),
             OrderClause {
-                field: withdrawal::properties::UPDATED_AT.to_string(),
+                field: withdrawal::properties::TRANSACTION_INDEX.to_string(),
                 ascending: true,
             },
         );
 
+        let cache = self.cache.read().unwrap();
+
+        let document_type = cache
+            .system_data_contracts
+            .withdrawals
+            .document_type_for_name(withdrawal::NAME)?;
+
         let drive_query = DriveQuery {
-            contract: &contract_fetch_info.contract,
+            contract: &cache.system_data_contracts.withdrawals,
             document_type,
             internal_clauses: InternalClauses {
                 primary_key_in_clause: None,
@@ -74,7 +78,7 @@ impl Drive {
                 equal_clauses: where_clauses,
             },
             offset: None,
-            limit: Some(100),
+            limit: Some(limit),
             order_by,
             start_at: None,
             start_at_included: false,
@@ -95,7 +99,8 @@ impl Drive {
 
 #[cfg(test)]
 mod tests {
-    use super::*;
+    use crate::drive::config::DEFAULT_QUERY_LIMIT;
+    use crate::drive::identity::withdrawals::WithdrawalTransactionIndex;
     use crate::tests::helpers::setup::{
         setup_document, setup_drive_with_initial_state_structure, setup_system_data_contract,
     };
@@ -109,8 +114,10 @@ mod tests {
     use dpp::version::PlatformVersion;
     use dpp::withdrawal::Pooling;
 
+    use super::*;
+
     #[test]
-    fn test_return_list_of_documents() {
+    fn test_find_pooled_withdrawal_documents_by_transaction_index() {
         let drive = setup_drive_with_initial_state_structure();
 
         let transaction = drive.grove.start_transaction();
@@ -123,17 +130,9 @@ mod tests {
 
         setup_system_data_contract(&drive, &data_contract, Some(&transaction));
 
-        let documents = drive
-            .fetch_up_to_100_oldest_withdrawal_documents_by_status(
-                withdrawals_contract::WithdrawalStatus::QUEUED.into(),
-                Some(&transaction),
-                platform_version,
-            )
-            .expect("to fetch documents by status");
-
-        assert_eq!(documents.len(), 0);
-
         let owner_id = Identifier::new([1u8; 32]);
+
+        let transaction_index: WithdrawalTransactionIndex = 1;
 
         let document = get_withdrawal_document_fixture(
             &data_contract,
@@ -143,13 +142,13 @@ mod tests {
                 "coreFeePerByte": 1u32,
                 "pooling": Pooling::Never as u8,
                 "outputScript": CoreScript::from_bytes((0..23).collect::<Vec<u8>>()),
-                "status": withdrawals_contract::WithdrawalStatus::QUEUED as u8,
-                "transactionIndex": 1u64,
+                "status": withdrawals_contract::WithdrawalStatus::POOLED as u8,
+                "transactionIndex": transaction_index,
             }),
             None,
             platform_version.protocol_version,
         )
-        .expect("expected withdrawal document");
+        .expect("expected to get withdrawal document");
 
         let document_type = data_contract
             .document_type_for_name(withdrawal::NAME)
@@ -163,48 +162,16 @@ mod tests {
             Some(&transaction),
         );
 
-        let document = get_withdrawal_document_fixture(
-            &data_contract,
-            owner_id,
-            platform_value!({
-                "amount": 1000u64,
-                "coreFeePerByte": 1u32,
-                "pooling": Pooling::Never as u8,
-                "outputScript": CoreScript::from_bytes((0..23).collect::<Vec<u8>>()),
-                "status": withdrawals_contract::WithdrawalStatus::POOLED,
-                "transactionIndex": 2u64,
-            }),
-            None,
-            platform_version.protocol_version,
-        )
-        .expect("expected withdrawal document");
-
-        setup_document(
-            &drive,
-            &document,
-            &data_contract,
-            document_type,
-            Some(&transaction),
-        );
-
-        let documents = drive
-            .fetch_up_to_100_oldest_withdrawal_documents_by_status(
-                withdrawals_contract::WithdrawalStatus::QUEUED.into(),
+        let found_document = drive
+            .find_withdrawal_documents_by_status_and_transaction_indices(
+                withdrawals_contract::WithdrawalStatus::POOLED,
+                &[transaction_index],
+                DEFAULT_QUERY_LIMIT,
                 Some(&transaction),
                 platform_version,
             )
-            .expect("to fetch documents by status");
+            .expect("to find document by it's transaction id");
 
-        assert_eq!(documents.len(), 1);
-
-        let documents = drive
-            .fetch_up_to_100_oldest_withdrawal_documents_by_status(
-                withdrawals_contract::WithdrawalStatus::POOLED.into(),
-                Some(&transaction),
-                platform_version,
-            )
-            .expect("to fetch documents by status");
-
-        assert_eq!(documents.len(), 1);
+        assert_eq!(found_document.len(), 1);
     }
 }
