+++ packages/rs-drive/src/drive/identity/withdrawals/document/mod.rs
@@ -1,4 +1,4 @@
 /// This module dedicated for a versioned fetch_up_to_100_oldest_withdrawal_documents_by_status
-pub mod fetch_up_to_100_oldest_withdrawal_documents_by_status;
+pub mod fetch_oldest_withdrawal_documents_by_status;
 /// This module dedicated for a versioned find_up_to_100_withdrawal_documents_by_status_and_transaction_indices
-pub mod find_up_to_100_withdrawal_documents_by_status_and_transaction_indices;
+pub mod find_withdrawal_documents_by_status_and_transaction_indices;
