+++ packages/rs-platform-version/src/version/v1.rs
@@ -364,7 +364,7 @@ pub(super) const PLATFORM_V1: PlatformVersion = PlatformVersion {
                 },
                 withdrawals: DriveIdentityWithdrawalMethodVersions {
                     document: DriveIdentityWithdrawalDocumentMethodVersions {
-                        fetch_up_to_100_oldest_withdrawal_documents_by_status: 0,
+                        fetch_oldest_withdrawal_documents_by_status: 0,
                         find_up_to_100_withdrawal_documents_by_status_and_transaction_indices: 0,
                     },
                     transaction: DriveIdentityWithdrawalTransactionMethodVersions {
